+++ core/src/main/java/com/nabla/sdk/core/data/auth/TokenLocalDataSource.kt
@@ -1,10 +1,12 @@
 package com.nabla.sdk.core.data.auth
 
+import com.nabla.sdk.core.domain.entity.AccessToken
 import com.nabla.sdk.core.domain.entity.AuthTokens
+import com.nabla.sdk.core.domain.entity.RefreshToken
 
 internal class TokenLocalDataSource {
-    private var refreshToken: String? = null
-    private var accessToken: String? = null
+    private var refreshToken: RefreshToken? = null
+    private var accessToken: AccessToken? = null
 
     fun setAuthTokens(authTokens: AuthTokens) {
         this.refreshToken = authTokens.refreshToken
@@ -15,7 +17,7 @@ internal class TokenLocalDataSource {
         val currentRefreshToken = refreshToken
         val currentAccessToken = accessToken
         return if (currentRefreshToken != null && currentAccessToken != null) {
-            AuthTokens(currentRefreshToken, currentAccessToken)
+            AuthTokens(currentAccessToken, currentRefreshToken)
         } else {
             null
         }