+++ messaging-sample-app/src/main/java/com/nabla/sdk/messagingsampleapp/MessagingSampleApp.kt
@@ -3,8 +3,10 @@ package com.nabla.sdk.messagingsampleapp
 import android.app.Application
 import android.util.Log
 import com.nabla.sdk.core.NablaClient
+import com.nabla.sdk.core.domain.entity.AccessToken
 import com.nabla.sdk.core.domain.entity.AuthTokens
 import com.nabla.sdk.core.domain.entity.AuthenticationException
+import com.nabla.sdk.core.domain.entity.RefreshToken
 import com.nabla.sdk.messaging.core.NablaMessagingModule
 import com.nabla.sdk.videocall.NablaVideoCallModule
 
@@ -18,8 +20,8 @@ internal class MessagingSampleApp : Application() {
                 // In your app, you need to replace this with an actual call to your backend to get fresh tokens
                 Result.success(
                     AuthTokens(
-                        refreshToken = "dummy-refresh-token",
-                        accessToken = "dummy-access-token",
+                        AccessToken("dummy-access-token"),
+                        RefreshToken("dummy-refresh-token"),
                     )
                 )
             }